requirements.txt
basicsr==1.4.2
imageio==2.35.1
matplotlib==3.7.5
numpy==1.24.4
opencv_contrib_python==4.12.0.88
opencv_python==4.10.0.82
opencv_python_headless==4.12.0.88
openpyxl==3.1.2
pandas==2.0.3
Pillow==12.0.0
piq==0.8.0
psnr_hvsm==0.2.4
pyrtools==1.0.10
pytorch_msssim==0.2.1
PyYAML==6.0.3
realesrgan==0.3.0
Requests==2.32.5
rich==14.2.0
scikit_learn==1.0.2
scipy==1.16.3
skimage==0.0
tools==1.0.13
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
webui==0.1.7
